requirements.txt
requests
nsepython
google-genai